Extract parameter resolution code from ExecutableInvoker
